+++ NEWS.md
@@ -1,3 +1,12 @@
+# fitur 0.6.1
+
+* shiny gadget for fitting univariate distributions has been added
+
+* added test function for distfun objects
+
+* diagnostic plots now have better checks for distfun objects and lists of
+distfun objects
+
 # fitur 0.6.0
 
 * Adding continuous distribution testing functions Kolmogorov-Smirnov, Anderson-Darling, and Cramer-Von Mises